+++ trustly-android-sdk/src/main/java/net/trustly/android/sdk/util/last_used/LastUsedBankManager.kt
@@ -1,4 +1,4 @@
-package net.trustly.android.sdk.util.manager
+package net.trustly.android.sdk.util.last_used
 
 import net.trustly.android.sdk.data.TrustlyStorage
 
@@ -15,6 +15,6 @@ class LastUsedBankManager(private val trustlyStorage: TrustlyStorage) {
         trustlyStorage.saveData(LAST_USED_BANK_ID, lastUsedBank)
     }
 
-    fun getLastUsedBank() = trustlyStorage.readStringDataFrom(LAST_USED_BANK_ID)
+    fun getLastUsedBank(): String? = trustlyStorage.readStringDataFrom(LAST_USED_BANK_ID)
 
 }
